toCpp: Split the consumer base by topic
vulkan_cpp_consumer_base.cpp held 3592 lines and mixed file handling,
platform output, name allocation, and the emitters for every API
call.

Move the emitters into nine files, one for each topic, and leave the
orchestration in the base file, which is now 876 lines. The header
does not change. The generated output is identical for all 14 test
captures.
